e2e-test/requirement.txt
Faker==8.0.0
requests==2.25.1
robotframework-selenium2library==3.0.0
selenium==3.141.0
robotframework==3.0.2
robotframework-faker==3.0.0